divide return types, choose double

+++ src/main/java/com/example/springdocker/MyMathCalc.java
@@ -16,7 +16,8 @@ public class MyMathCalc {
         return a * b;
     }
 
-    public float divide(int a, int b){
+
+    public double divide(int a, int b){
         return a/b;
     }
 }
